Balance charge implementation: BE side
