+++ packages/server/src/modules/charges/resolvers/charges.resolver.ts
@@ -4,10 +4,8 @@ import { DocumentsProvider } from '@modules/documents/providers/documents.provid
 import { ledgerGenerationByCharge } from '@modules/ledger/helpers/ledger-by-charge-type.helper.js';
 import { ledgerRecordsGenerationFullMatchComparison } from '@modules/ledger/helpers/ledgrer-storage.helper.js';
 import { LedgerProvider } from '@modules/ledger/providers/ledger.provider.js';
-import { generateLedgerRecordsForFinancialCharge } from '@modules/ledger/resolvers/ledger-generation/financial-ledger-generation.resolver.js';
 import { MiscExpensesProvider } from '@modules/misc-expenses/providers/misc-expenses.provider.js';
 import { ChargeTagsProvider } from '@modules/tags/providers/charge-tags.provider.js';
-import { TagsProvider } from '@modules/tags/providers/tags.provider.js';
 import { TransactionsProvider } from '@modules/transactions/providers/transactions.provider.js';
 import { EMPTY_UUID } from '@shared/constants';
 import { ChargeSortByField, ChargeTypeEnum } from '@shared/enums';
@@ -409,346 +407,6 @@ export const chargesResolvers: ChargesModule.Resolvers &
         throw new GraphQLError(`Error deleting charge ID="${chargeId}"`);
       }
     },
-    generateRevaluationCharge: async (_, { date, ownerId }, context, info) => {
-      const { injector, adminContext } = context;
-      try {
-        const [charge] = await injector.get(ChargesProvider).generateCharge({
-          ownerId,
-          userDescription: `Revaluation charge for ${date}`,
-          type: 'FINANCIAL',
-          taxCategoryId: adminContext.general.taxCategories.exchangeRevaluationTaxCategoryId,
-        });
-
-        if (!charge) {
-          throw new Error('Error creating new charge');
-        }
-
-        const newExtendedCharge = await injector
-          .get(ChargesProvider)
-          .getChargeByIdLoader.load(charge.id);
-
-        if (!newExtendedCharge) {
-          throw new Error('Error creating new charge');
-        }
-
-        await generateLedgerRecordsForFinancialCharge(
-          newExtendedCharge,
-          { insertLedgerRecordsIfNotExists: true },
-          context,
-          info,
-        );
-
-        return newExtendedCharge;
-      } catch (e) {
-        console.error(e);
-        throw new GraphQLError('Error generating revaluation charge');
-      }
-    },
-    generateBankDepositsRevaluationCharge: async (_, { date, ownerId }, context, info) => {
-      const { injector, adminContext } = context;
-      try {
-        const { bankDepositInterestIncomeTaxCategoryId } = adminContext.bankDeposits;
-        if (!bankDepositInterestIncomeTaxCategoryId) {
-          throw new GraphQLError('Bank deposit interest income tax category missing');
-        }
-        const [charge] = await injector.get(ChargesProvider).generateCharge({
-          ownerId,
-          userDescription: `Bank deposits revaluation charge for ${date}`,
-          type: 'FINANCIAL',
-          taxCategoryId: bankDepositInterestIncomeTaxCategoryId,
-        });
-
-        if (!charge) {
-          throw new Error('Error creating new charge');
-        }
-
-        const newExtendedCharge = await injector
-          .get(ChargesProvider)
-          .getChargeByIdLoader.load(charge.id);
-
-        if (!newExtendedCharge) {
-          throw new Error('Error creating new charge');
-        }
-
-        await generateLedgerRecordsForFinancialCharge(
-          newExtendedCharge,
-          { insertLedgerRecordsIfNotExists: true },
-          context,
-          info,
-        );
-
-        return newExtendedCharge;
-      } catch (e) {
-        console.error(e);
-        throw new GraphQLError('Error generating bank deposits revaluation charge');
-      }
-    },
-    generateTaxExpensesCharge: async (_, { year, ownerId }, context, info) => {
-      const {
-        injector,
-        adminContext: {
-          authorities: { taxExpensesTaxCategoryId },
-        },
-      } = context;
-      try {
-        const [charge] = await injector.get(ChargesProvider).generateCharge({
-          ownerId,
-          userDescription: `Tax expenses charge for ${year.substring(0, 4)}`,
-          type: 'FINANCIAL',
-          taxCategoryId: taxExpensesTaxCategoryId,
-        });
-
-        if (!charge) {
-          throw new Error('Error creating new charge');
-        }
-
-        const newExtendedCharge = await injector
-          .get(ChargesProvider)
-          .getChargeByIdLoader.load(charge.id);
-
-        if (!newExtendedCharge) {
-          throw new Error('Error creating new charge');
-        }
-
-        const tagName = 'financial';
-
-        const addTagPromise = async () => {
-          const tag = await injector
-            .get(TagsProvider)
-            .getTagByNameLoader.load(tagName)
-            .catch(() => {
-              throw new GraphQLError(`Error adding "${tagName}" tag`);
-            });
-
-          if (!tag) {
-            throw new GraphQLError(`"${tagName}" tag not found`);
-          }
-
-          await injector
-            .get(ChargeTagsProvider)
-            .insertChargeTag({ chargeId: newExtendedCharge.id, tagId: tag.id })
-            .catch(() => {
-              throw new GraphQLError(
-                `Error adding "${tagName}" tag to charge ID="${newExtendedCharge.id}"`,
-              );
-            });
-        };
-
-        const generateLedgerPromise = generateLedgerRecordsForFinancialCharge(
-          newExtendedCharge,
-          { insertLedgerRecordsIfNotExists: true },
-          context,
-          info,
-        );
-
-        await Promise.all([addTagPromise(), generateLedgerPromise]);
-
-        return newExtendedCharge;
-      } catch (e) {
-        console.error(e);
-        throw new GraphQLError('Error generating tax expenses charge');
-      }
-    },
-    generateDepreciationCharge: async (_, { year, ownerId }, context, info) => {
-      const {
-        injector,
-        adminContext: {
-          depreciation: { accumulatedDepreciationTaxCategoryId },
-        },
-      } = context;
-      if (!accumulatedDepreciationTaxCategoryId) {
-        throw new GraphQLError('Accumulated depreciation tax category missing');
-      }
-      try {
-        const [charge] = await injector.get(ChargesProvider).generateCharge({
-          ownerId,
-          userDescription: `Depreciation charge for ${year.substring(0, 4)}`,
-          type: 'FINANCIAL',
-          taxCategoryId: accumulatedDepreciationTaxCategoryId,
-        });
-
-        if (!charge) {
-          throw new Error('Error creating new charge');
-        }
-
-        const newExtendedCharge = await injector
-          .get(ChargesProvider)
-          .getChargeByIdLoader.load(charge.id);
-
-        if (!newExtendedCharge) {
-          throw new Error('Error creating new charge');
-        }
-
-        const tagName = 'financial';
-
-        const addTagPromise = async () => {
-          const tag = await injector
-            .get(TagsProvider)
-            .getTagByNameLoader.load(tagName)
-            .catch(() => {
-              throw new GraphQLError(`Error adding "${tagName}" tag`);
-            });
-
-          if (!tag) {
-            throw new GraphQLError(`"${tagName}" tag not found`);
-          }
-
-          await injector
-            .get(ChargeTagsProvider)
-            .insertChargeTag({ chargeId: newExtendedCharge.id, tagId: tag.id })
-            .catch(() => {
-              throw new GraphQLError(
-                `Error adding "${tagName}" tag to charge ID="${newExtendedCharge.id}"`,
-              );
-            });
-        };
-
-        const generateLedgerPromise = generateLedgerRecordsForFinancialCharge(
-          newExtendedCharge,
-          { insertLedgerRecordsIfNotExists: true },
-          context,
-          info,
-        );
-
-        await Promise.all([addTagPromise(), generateLedgerPromise]);
-
-        return newExtendedCharge;
-      } catch (e) {
-        console.error(e);
-        throw new GraphQLError('Error generating depreciation charge');
-      }
-    },
-    generateRecoveryReserveCharge: async (_, { year, ownerId }, context, info) => {
-      const { injector, adminContext } = context;
-      const { recoveryReserveTaxCategoryId } = adminContext.salaries;
-      if (!recoveryReserveTaxCategoryId) {
-        throw new GraphQLError('Recovery reserve tax category missing');
-      }
-      try {
-        const [charge] = await injector.get(ChargesProvider).generateCharge({
-          ownerId,
-          userDescription: `Recovery reserve charge for ${year.substring(0, 4)}`,
-          type: 'FINANCIAL',
-          taxCategoryId: recoveryReserveTaxCategoryId,
-        });
-
-        if (!charge) {
-          throw new Error('Error creating new charge');
-        }
-
-        const newExtendedCharge = await injector
-          .get(ChargesProvider)
-          .getChargeByIdLoader.load(charge.id);
-
-        if (!newExtendedCharge) {
-          throw new Error('Error creating new charge');
-        }
-
-        const tagName = 'financial';
-
-        const addTagPromise = async () => {
-          const tag = await injector
-            .get(TagsProvider)
-            .getTagByNameLoader.load(tagName)
-            .catch(() => {
-              throw new GraphQLError(`Error adding "${tagName}" tag`);
-            });
-
-          if (!tag) {
-            throw new GraphQLError(`"${tagName}" tag not found`);
-          }
-
-          await injector
-            .get(ChargeTagsProvider)
-            .insertChargeTag({ chargeId: newExtendedCharge.id, tagId: tag.id })
-            .catch(() => {
-              throw new GraphQLError(
-                `Error adding "${tagName}" tag to charge ID="${newExtendedCharge.id}"`,
-              );
-            });
-        };
-
-        const generateLedgerPromise = generateLedgerRecordsForFinancialCharge(
-          newExtendedCharge,
-          { insertLedgerRecordsIfNotExists: true },
-          context,
-          info,
-        );
-
-        await Promise.all([addTagPromise(), generateLedgerPromise]);
-
-        return newExtendedCharge;
-      } catch (e) {
-        console.error(e);
-        throw new GraphQLError('Error generating recovery reserve charge');
-      }
-    },
-    generateVacationReserveCharge: async (_, { year, ownerId }, context, info) => {
-      const { injector, adminContext } = context;
-      const { vacationReserveTaxCategoryId } = adminContext.salaries;
-      if (!vacationReserveTaxCategoryId) {
-        throw new GraphQLError('Vacation reserve tax category missing');
-      }
-
-      try {
-        const [charge] = await injector.get(ChargesProvider).generateCharge({
-          ownerId,
-          userDescription: `Vacation reserves charge for ${year.substring(0, 4)}`,
-          type: 'FINANCIAL',
-          taxCategoryId: vacationReserveTaxCategoryId,
-        });
-
-        if (!charge) {
-          throw new Error('Error creating new charge');
-        }
-
-        const newExtendedCharge = await injector
-          .get(ChargesProvider)
-          .getChargeByIdLoader.load(charge.id);
-
-        if (!newExtendedCharge) {
-          throw new Error('Error creating new charge');
-        }
-
-        const tagName = 'financial';
-
-        const addTagPromise = async () => {
-          const tag = await injector
-            .get(TagsProvider)
-            .getTagByNameLoader.load(tagName)
-            .catch(() => {
-              throw new GraphQLError(`Error adding "${tagName}" tag`);
-            });
-
-          if (!tag) {
-            throw new GraphQLError(`"${tagName}" tag not found`);
-          }
-
-          await injector
-            .get(ChargeTagsProvider)
-            .insertChargeTag({ chargeId: newExtendedCharge.id, tagId: tag.id })
-            .catch(() => {
-              throw new GraphQLError(
-                `Error adding "${tagName}" tag to charge ID="${newExtendedCharge.id}"`,
-              );
-            });
-        };
-
-        const generateLedgerPromise = generateLedgerRecordsForFinancialCharge(
-          newExtendedCharge,
-          { insertLedgerRecordsIfNotExists: true },
-          context,
-          info,
-        );
-
-        await Promise.all([addTagPromise(), generateLedgerPromise]);
-
-        return newExtendedCharge;
-      } catch (e) {
-        console.error(e);
-        throw new GraphQLError('Error generating vacation reserves charge');
-      }
-    },
   },
   UpdateChargeResult: {
     __resolveType: (obj, _context, _info) => {
@@ -767,23 +425,6 @@ export const chargesResolvers: ChargesModule.Resolvers &
   CommonCharge: {
     __isTypeOf: (DbCharge, context) => getChargeType(DbCharge, context) === ChargeTypeEnum.Common,
     ...commonChargeFields,
-  },
-  FinancialCharge: {
-    __isTypeOf: (DbCharge, context) =>
-      getChargeType(DbCharge, context) === ChargeTypeEnum.Financial,
-    ...commonChargeFields,
-    vat: () => null,
-    totalAmount: () => null,
-    property: () => false,
-    conversion: () => false,
-    salary: () => false,
-    isInvoicePaymentDifferentCurrency: () => false,
-    minEventDate: DbCharge => DbCharge.ledger_min_invoice_date,
-    minDebitDate: DbCharge => DbCharge.ledger_min_value_date,
-    // minDocumentsDate:
-    // validationData:
-    // metadata:
-    yearsOfRelevance: () => null,
   },
   ConversionCharge: {
     __isTypeOf: (DbCharge, context) =>
